Implemented authorization & some bug fixes
- Added ignore words to cSpell
- Fixed the recaptchav3 bug where it does not verify unless the recaptchav3 code is anchored right before the closing of `<form>` tag
- Added some guards to prevent contradiction like a logged in user is trying to register
- Added `ErrorHandler` to handle errors that exists in the form to prevent form from being submitted and also used for logging
- Added authorization guard using function `universal_get_current_user_role()` for pages that should only be accessible by certain roles (incomplete as manager and admins should not be able to add booking)
- Added logging for all critical routes
- Removed old route `/admin`
- Updated all basic error pages that is of 400/500 series to flash messages
- Updated error messages that are logged to be more descriptive of the error
- Updated `/search` to use GET instead
- Updated regex patterns for input validation (provided by YJ)
- Added the preferred date format for this web app as a constant
- Added checks to prevent users from deleting/canceling their booking once the confirm has been completed or canceled
- Removed unused method `validate_str_input()` on `input_validation`
- Updated missing checks in `validate_image()`. The image may not always have an extension (TODO: get only the last `.` character and assume everything to the left is the file extension)
- Added new validation function `validate_date()`

+++ flasks/bp_fcp.py
@@ -1,24 +1,62 @@
-from flask import Blueprint, request, redirect, url_for, render_template, flash, abort
+from flask import Blueprint, request, redirect, url_for, render_template, flash, abort, current_app
+import flask_login
 
 from db import db
 
 from fault import Fault
 from booking import Booking
+from user import Role
+
+from authorizer import universal_get_current_user_role
+from error_handler import ErrorHandler
 
 bp_fcp = Blueprint('bp_fcp', __name__, template_folder='templates')
 
 
 @bp_fcp.route("/manager/fcp", methods=["GET"])
 def manager_read_faults() -> str:
-    faults = Fault.query.all()
-    bookings = Booking.query.all()
+    if universal_get_current_user_role(flask_login.current_user) == Role.MANAGER:
+        faults = Fault.query.all()
+        bookings = Booking.query.all()
 
-    return render_template("manager-faults.html", booking_list=bookings, fault_list=faults)
+        return render_template("manager-faults.html", booking_list=bookings, fault_list=faults)
+    else:
+        err_handler = ErrorHandler(current_app, dict(request.headers))
+        user_email = "Anonymous" if universal_get_current_user_role(flask_login.current_user) == 0 else flask_login.current_user.email
+
+        err_handler.push(
+            user_message="",
+            log_message=f"Unauthorized access from user {user_email}"
+        )
+
+        err_handler.commit_log()
+
+        abort(401)
 
 
 @bp_fcp.route("/manager/fcp/fault/create", methods=["POST"])
 def manager_create_fault() -> str:
-    return abort(501, "This should never be used?")
+    err_handler = ErrorHandler(current_app, dict(request.headers))
+    user_email = "Anonymous" if universal_get_current_user_role(flask_login.current_user) == 0 else flask_login.current_user.email
+
+    if universal_get_current_user_role(flask_login.current_user) == Role.MANAGER:
+        err_handler.push(
+            user_message="This should never be used?",
+            log_message=f"User {user_email} accessed a route that is not implemented."
+        )
+
+        err_handler.commit_log()
+
+        return abort(501, err_handler.first().user_message)
+    else:
+        err_handler.push(
+            user_message="",
+            log_message=f"Unauthorized access from user {user_email} to a route that is not implemented"
+        )
+
+        err_handler.commit_log()
+
+        abort(401)
     # booking_id = request.form.get("booking_id")
     # reported_date = request.form.get("reported_date")
     # description = request.form.get("description")
@@ -43,34 +81,91 @@ def manager_read_fault(fault_id: int) -> str:
     # bookings = Booking.query.all()
 
     # return redirect(url_for("manager_read_faults"))
-    return abort(501, "This should never be used?")
+    err_handler = ErrorHandler(current_app, dict(request.headers))
+    user_email = "Anonymous" if universal_get_current_user_role(flask_login.current_user) == 0 else flask_login.current_user.email
+
+    if universal_get_current_user_role(flask_login.current_user) == Role.MANAGER:
+        err_handler.push(
+            user_message="This should never be used?",
+            log_message=f"User {user_email} accessed a route that is not implemented."
+        )
+
+        err_handler.commit_log()
+
+        return abort(501, err_handler.first().user_message)
+    else:
+        err_handler.push(
+            user_message="",
+            log_message=f"Unauthorized access from user {user_email} to a route that is not implemented"
+        )
+
+        err_handler.commit_log()
+
+        abort(401)
 
 
 @bp_fcp.route("/manager/fcp/fault/update/<int:fault_id>", methods=["POST"])
 def manager_update_fault(fault_id: int) -> str:
-    booking_id = request.form.get("booking_id")
-    reported_date = request.form.get("reported_date")
-    description = request.form.get("description")
+    err_handler = ErrorHandler(current_app, dict(request.headers))
+    user_email = "Anonymous" if universal_get_current_user_role(flask_login.current_user) == 0 else flask_login.current_user.email
 
-    update_dict = {
-        "booking_id": booking_id,
-        "reported_date": reported_date,
-        "description": description
-    }
+    if universal_get_current_user_role(flask_login.current_user) == Role.MANAGER:
+        booking_id = request.form.get("booking_id")
+        reported_date = request.form.get("reported_date")
+        description = request.form.get("description")
 
-    update_dict = {k: v for k, v in update_dict.items() if v is not None}
+        update_dict = {
+            "booking_id": booking_id,
+            "reported_date": reported_date,
+            "description": description
+        }
 
-    Fault.query.filter_by(fault_id=fault_id).update(update_dict)
-    db.session.commit()
+        update_dict = {k: v for k, v in update_dict.items() if v is not None}
 
-    flash("Fault updated!", category="success")
-    return redirect(url_for("bp_fcp.manager_read_faults"))
+        Fault.query.filter_by(fault_id=fault_id).update(update_dict)
+        db.session.commit()
+
+        flash("Fault updated!", category="success")
+
+        err_handler.push(
+            user_message="",
+            log_message=f"Fault ID '{fault_id}' has been updated. Request made by user {user_email}"
+        )
+
+        err_handler.commit_log()
+        return redirect(url_for("bp_fcp.manager_read_faults"))
+    else:
+        err_handler.push(
+            user_message="",
+            log_message=f"Unauthorized access from user {user_email}"
+        )
+
+        err_handler.commit_log()
+
+        abort(401)
 
 
 @bp_fcp.route("/manager/fcp/fault/delete/<int:fault_id>", methods=["POST"])
 def manager_delete_fault(fault_id: int) -> str:
-    Fault.query.filter_by(fault_id=fault_id).delete()
-    db.session.commit()
+    err_handler = ErrorHandler(current_app, dict(request.headers))
+    user_email = "Anonymous" if universal_get_current_user_role(flask_login.current_user) == 0 else flask_login.current_user.email
 
-    flash("Fault deleted!", category="success")
-    return redirect(url_for("bp_fcp.manager_read_faults"))
+    if universal_get_current_user_role(flask_login.current_user) == Role.MANAGER:
+        Fault.query.filter_by(fault_id=fault_id).delete()
+        db.session.commit()
+
+        flash("Fault deleted!", category="success")
+        err_handler.push(
+            user_message="",
+            log_message=f"Fault ID '{fault_id}' has been deleted. Request made by user {user_email}"
+        )
+        return redirect(url_for("bp_fcp.manager_read_faults"))
+    else:
+        err_handler.push(
+            user_message="",
+            log_message=f"Unauthorized access from user {user_email}"
+        )
+
+        err_handler.commit_log()
+
+        abort(401)
